+++ app/package.json
@@ -99,7 +99,7 @@
     "papaparse": "^5.4.1",
     "phosphor-react": "^1.3.1",
     "postcss-flexbugs-fixes": "^5.0.2",
-    "posthog-js": "^1.150.0",
+    "posthog-js": "^1.161.3",
     "prettier": "^2.3.1",
     "re-resizable": "^6.9.0",
     "react": "^18.2.0",
