+++ app/src/components/Layout.tsx
@@ -1,6 +1,6 @@
 import cx from "classnames";
 import { X } from "phosphor-react";
-import { memo, ReactNode, Suspense } from "react";
+import { lazy, memo, ReactNode, Suspense } from "react";
 import { Link } from "react-router-dom";
 
 import { useFullscreen, useIsEditorView } from "../lib/hooks";
@@ -10,7 +10,7 @@ import { Header } from "./Header";
 import styles from "./Layout.module.css";
 import Loading from "./Loading";
 import { VersionCheck } from "./VersionCheck";
-import { PaywallModal } from "./PaywallModal";
+const PaywallModal = lazy(() => import("./PaywallModal"));
 
 const Layout = memo(({ children }: { children: ReactNode }) => {
   const isFullscreen = useFullscreen();
@@ -49,7 +49,9 @@ const Layout = memo(({ children }: { children: ReactNode }) => {
         <ColorMode />
       </Box>
       <VersionCheck />
-      <PaywallModal />
+      <Suspense fallback={null}>
+        <PaywallModal />
+      </Suspense>
     </>
   );
 });
